refactor: Question 도메인을 Template 아래로 이동
- 전체적인 경계를 아래와 같이 나눔:
1. 템플릿 (질문, 옵션, 섹션, 템플릿)
2. 리뷰 (답변)
3. 하이라이트
4. 리뷰그
