+++ backend/src/main/java/reviewme/DatabaseInitializer.java
@@ -5,14 +5,14 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
-import reviewme.question.domain.OptionGroup;
-import reviewme.question.domain.OptionItem;
-import reviewme.question.domain.OptionType;
-import reviewme.question.domain.Question;
-import reviewme.question.domain.QuestionType;
-import reviewme.question.repository.OptionGroupRepository;
-import reviewme.question.repository.OptionItemRepository;
-import reviewme.question.repository.QuestionRepository;
+import reviewme.template.domain.OptionGroup;
+import reviewme.template.domain.OptionItem;
+import reviewme.template.domain.OptionType;
+import reviewme.template.domain.Question;
+import reviewme.template.domain.QuestionType;
+import reviewme.template.repository.OptionGroupRepository;
+import reviewme.template.repository.OptionItemRepository;
+import reviewme.template.repository.QuestionRepository;
 import reviewme.template.domain.Section;
 import reviewme.template.domain.Template;
 import reviewme.template.domain.VisibleType;
