+++ backend/src/main/java/reviewme/review/service/dto/response/gathered/SimpleQuestionResponse.java
@@ -1,6 +1,6 @@
 package reviewme.review.service.dto.response.gathered;
 
-import reviewme.question.domain.QuestionType;
+import reviewme.template.domain.QuestionType;
 
 public record SimpleQuestionResponse(
         long id,
