+++ backend/src/main/java/reviewme/review/service/mapper/AnswerMapper.java
@@ -1,6 +1,6 @@
 package reviewme.review.service.mapper;
 
-import reviewme.question.domain.QuestionType;
+import reviewme.template.domain.QuestionType;
 import reviewme.review.domain.Answer;
 import reviewme.review.service.dto.request.ReviewAnswerRequest;
