+++ backend/src/main/java/reviewme/review/service/mapper/ReviewMapper.java
@@ -8,8 +8,8 @@ import java.util.stream.Collectors;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import reviewme.question.domain.Question;
-import reviewme.question.repository.QuestionRepository;
+import reviewme.template.domain.Question;
+import reviewme.template.repository.QuestionRepository;
 import reviewme.review.domain.Answer;
 import reviewme.review.domain.Review;
 import reviewme.review.service.dto.request.ReviewAnswerRequest;
