+++ backend/src/main/java/reviewme/template/domain/OptionItem.java
@@ -1,4 +1,4 @@
-package reviewme.question.domain;
+package reviewme.template.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
