+++ backend/src/main/java/reviewme/template/repository/OptionItemRepository.java
@@ -1,11 +1,11 @@
-package reviewme.question.repository;
+package reviewme.template.repository;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import reviewme.question.domain.OptionItem;
-import reviewme.question.domain.OptionType;
+import reviewme.template.domain.OptionItem;
+import reviewme.template.domain.OptionType;
 
 @Repository
 public interface OptionItemRepository extends JpaRepository<OptionItem, Long> {
