+++ backend/src/main/java/reviewme/template/repository/QuestionRepository.java
@@ -1,12 +1,12 @@
-package reviewme.question.repository;
+package reviewme.template.repository;
 
 import java.util.List;
 import java.util.Set;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import reviewme.question.domain.OptionItem;
-import reviewme.question.domain.Question;
+import reviewme.template.domain.OptionItem;
+import reviewme.template.domain.Question;
 
 @Repository
 public interface QuestionRepository extends JpaRepository<Question, Long> {
