+++ backend/src/test/java/reviewme/review/repository/ReviewRepositoryTest.java
@@ -13,8 +13,8 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import reviewme.question.domain.Question;
-import reviewme.question.repository.QuestionRepository;
+import reviewme.template.domain.Question;
+import reviewme.template.repository.QuestionRepository;
 import reviewme.review.domain.Review;
 import reviewme.reviewgroup.domain.ReviewGroup;
 import reviewme.reviewgroup.repository.ReviewGroupRepository;
