+++ backend/src/test/java/reviewme/review/service/ReviewDetailLookupServiceTest.java
@@ -17,12 +17,12 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import reviewme.question.domain.OptionGroup;
-import reviewme.question.domain.OptionItem;
-import reviewme.question.domain.Question;
-import reviewme.question.repository.OptionGroupRepository;
-import reviewme.question.repository.OptionItemRepository;
-import reviewme.question.repository.QuestionRepository;
+import reviewme.template.domain.OptionGroup;
+import reviewme.template.domain.OptionItem;
+import reviewme.template.domain.Question;
+import reviewme.template.repository.OptionGroupRepository;
+import reviewme.template.repository.OptionItemRepository;
+import reviewme.template.repository.QuestionRepository;
 import reviewme.review.domain.Answer;
 import reviewme.review.domain.CheckboxAnswer;
 import reviewme.review.domain.Review;
