+++ backend/src/test/java/reviewme/review/service/validator/TextAnswerValidatorTest.java
@@ -9,8 +9,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
-import reviewme.question.domain.Question;
-import reviewme.question.repository.QuestionRepository;
+import reviewme.template.domain.Question;
+import reviewme.template.repository.QuestionRepository;
 import reviewme.review.domain.TextAnswer;
 import reviewme.review.service.exception.InvalidTextAnswerLengthException;
 import reviewme.review.service.exception.SubmittedQuestionNotFoundException;
