+++ backend/src/test/java/reviewme/template/repository/OptionGroupRepositoryTest.java
@@ -1,7 +1,6 @@
-package reviewme.question.repository;
+package reviewme.template.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 import static reviewme.fixture.OptionGroupFixture.선택지_그룹;
 import static reviewme.fixture.QuestionFixture.선택형_필수_질문;
 
@@ -9,8 +8,10 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import reviewme.question.domain.OptionGroup;
-import reviewme.question.domain.Question;
+import reviewme.template.domain.OptionGroup;
+import reviewme.template.domain.Question;
+import reviewme.template.repository.OptionGroupRepository;
+import reviewme.template.repository.QuestionRepository;
 
 @DataJpaTest
 class OptionGroupRepositoryTest {
